backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pyyaml>=6.0
